refactor(α): wire-type foundation — single source of truth for Backend* shapes
First slice of the architecture refactor (phase α). Establishes the
foundation pieces that phases β and δ build on; no behavior change.

What's new
* `worker/types/wire/{matter,invoice}.ts` — backend HTTP wire types live
  here, snake_case, exactly matching `BACKEND_API_URL`. New code imports
  via `@/shared/types/wire`; matter/invoice are migrated as the pilot.
* `src/shared/types/wire.ts` — frontend re-export shim. Single import
  path for any backend response/request shape.
* `unwrapApiResponse<T>(payload)` + `pluckCollection<T>(value, keys)` in
  `src/shared/lib/apiClient.ts` — replaces the inline
  `.data ?? .matters ?? .items ?? recurse` patterns each service
  reinvented (`extractMatterArray`, `extractNotesArray`, `requestData`,
  …). Throws on the `success: false` branch so callers don't have to
  hand-write that check.
* `ApiResponse<T>` in `worker/types.ts` is now a discriminated union
  (`{ success: true; data } | { success: false; error; … }`) so
  consumers narrow on `success` instead of runtime-checking every field.
  No producer changes needed — `createSuccessResponse` and `handleError`
  in `worker/errorHandler.ts` already emit shapes the union accepts.

Migrated
* `src/features/matters/services/mattersApi.ts` — the 8 `Backend*` types
  (Matter/Activity/Note/TimeEntry/TimeStats/Expense/Milestone/Task plus
  TaskStatus/TaskPriority enums) are now re-exported from the wire
  module instead of declared inline. The 6 consumers that import
  `BackendMatter` from `mattersApi` still work.
* `src/features/matters/services/invoicesApi.ts` — same treatment for
  `BackendInvoice` / `BackendInvoiceLineItem`.

Out of scope for this slice (queued in α task)
* Migrating ChatMessage/ChatSession/Practice/Workspace from worker/types.ts
  into wire/ (mostly cosmetic split).
* Adopting `unwrapApiResponse`/`pluckCollection` at the existing call
  sites (`extractMatterArray` and friends still inline).
* ESLint rule banning `Backend*` declarations outside `worker/types/`.
* Zod runtime schemas for wire types (cheap validation at the
  worker/frontend boundary).

Verification
* npx tsc --noEmit: 0 errors
* npm run test:unit: 243/243 pass
* npm run test:component: 75/75 pass
